requirements.txt
Django==5.0.7
djangorestframework==3.15.2
Pillow==10.4.0
pandas==2.3.1
numpy==2.3.2
opencv-python-headless==4.10.0.84
google-generativeai==0.7.2
python-multipart==0.0.9
whitenoise==6.6.0
